calculate left for poker deck
